requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
jinja2>=3.1.0,<4.0.0
httpx>=0.27.0,<1.0.0
pydantic>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
